pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dex-adapter"
version = "1.1.0"
description = "DEX Adapter - unified interface for Solana (Raydium, Meteora, Jupiter) and EVM (1inch) protocols"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "solana>=0.30.0",
    "solders>=0.20.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# EVM chain support (Ethereum/BSC via 1inch)
evm = [
    "web3>=6.0.0",
    "eth-account>=0.8.0",
]
# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
# All optional dependencies
all = [
    "web3>=6.0.0",
    "eth-account>=0.8.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["dex_adapter*"]
